xen-irq-latency: switch DT ovelays to be compatible with hello-mod
For hello-mod we want to build for xl and for dom0less.
Lets do the same here.

This makes the build command line for the normal case of dom0less more
awkward but allows us to make an xl version.  This may not make sense for
this app but it could be of use in testing.

Signed-off-by: Bill Mills <bill.mills@linaro.org>